+++ express/src/decorators/controller.ts
@@ -1,13 +1,13 @@
 import { ExpressMeta, getMeta } from '../meta';
-import { Middleware } from '../middleware';
+import { Type } from '../middleware';
 
 /**
  * Registers controller for base url
  *
  * @param {string} url
- * @param {Middleware[]} [middleware]
+ * @param {Type[]} [middleware]
  */
-export function Controller(url: string, middleware?: Middleware[]) {
+export function Controller(url: string, middleware?: Type[]) {
   return (target): void => {
     const meta: ExpressMeta = getMeta(target.prototype);